*Font: removed doOpenFile() implementation.
AbstractFont does that internally and IMHO the code duplication is not
worth possible performance/memory improvements when opening the file
directly with FreeType (which mmaps it, probably) instead of loading it
into memory and then opening the memory and keeping the memory for whole
FreeType lifetime.

Also added an assertion for internal state, which exposed (probably
harmless) bug on closing.