OpenGexImporter: selecting structures from a set of identifiers.
This is very ugly, because as it currently is, both findFirstChildOf()
and findNextOf() take std::initializer_list as argument (thus {}), but
childrenOf() unintuively takes variadic argument list.

That is because with std::initializer_list I was not able to keep it
in scope for the whole lifetime of range-based for loop -- it became
random memory right after exiting the childrenOf() function. One
solution was to copy the initializer_list to heap-allocated array, but
that's one allocation per each for loop, which is insane. This solution
creates compile-time-sized array on stack, which is slightly better, but
causes code bloat (templatws...) and has the unintuitive behavior
mentioned above. If someone could come up with better solution (perhaps
making use of constexpr std::intializer_list::size() in C++14), I'm all
ears.