OpenGexImporter: fix garbage memory access in Document::validate().
Copying std::initializer_list into variable and then expecting that it
would be valid even after the original list goes out of scope was a
*bad* idea. Apparently that worked only GCC 4.9, the code goes haywire
everywhere else.

Because of that, the validation tree now needs to allocate its own
arrays on heap and copy the contents of the lists to them. It's uglier,
slower, not memory efficient and all, but I could not think of any
better solution.

Also, because Corrade::Containers::Array doesn't *yet* have ability to
allocate uninitialized memory, I had to add hacky "internal" default
constructor. The proper solution would take more time, but I now need to
push the release out of the door.