DdsImporter: give up on simplicity and use strongly typed enums and EnumSet.
The problem was the DdsDescriptionFlag::PixelFormat enum value. Because
it was not a strongly typed enum and I have recently renamed ColorFormat
to PixelFormat, this was causing GCC to break in completely crazy way and
I had to fire up Clang to get some proper error message.

The path of least resistance would be to use scoped enums that are
implicitly convertible to UnsignedInt, but hey, that's not possible in
C++. So I have to use strongly typed enums, EnumSet and change struct
members from UnsignedInt to EnumSet instances. Added a sanity
static_assert() to verify that the struct still has proper size.

At least the enum values won't now collide with random names.