CMake subproject support, part 2: provide MAGNUMPLUGINS_INCLUDE_DIRS var.
When using plugins repo as subprojects, its root include dir is
different from Magnum include dir, thus the user needs to include both.
This variable was accidentally mentioned in docs even though it didn't
exist. So it exists now and everyone should have been using it all
along ;)

Also use HINTS instead of PATHS for find_path() because the value is not
hardcoded but rather depending on previous introspection and CMake
documentation suggests using different keyword for this.