Modern CMake usage: reworked to use target-specific properties.
Please note that this is not exactly backwards compatible, but as most
uses of the plugin library are without interacting with CMake, it
shouldn't be as painful, hopefully. In any case, recreate the build dir
and update your copy of Find*.cmake modules to avoid weird things
happening.

User-facing changes:

 * Documentation of Find*.cmake modules converted to reStructuredText to
   follow official CMake guidelines.
 * The newfangled way to link the plugins is to use
   MagnumPlugins::JpegImporter instead of
   ${MAGNUMPLUGINS_JPEGIMPORTER_LIBRARIES} and
   ${MAGNUMPLUGINS_JPEGIMPORTER_INCLUDE_DIRS}.
 * The old ${MAGNUMPLUGINS_*_LIBRARIES} are deprecated and
   just expand to MagnumPlugins::* target and are exposed only if
   building with BUILD_DEPRECATED.
 * The old ${MAGNUMPLUGINS_*_INCLUDE_DIRS} are removed, as the
   MagnumPlugins::* targets cover that now as well.

Internal changes:

 * Removed global state in favor of target-specific settings.