modules: make FindDevIL actually fail when the something isn't found.
Due to a mismatch between module name and name passed to FPHSA() the
macro reported a failure, but that was not picked up by CMake so it
silently passed, failing much later during compilation. Did anyone
actually test the file?!

The ArchLinux distribution of DevIL doesn't contain ILUT, so the module
fails. We don't use that library, so I just removed it from the find
module altogether.