PngImporter, PngImageConverter: assert that there is no DLL mismatch.
Commiter note (mosra): this was originally a runtime error, but I think,
since there is nothing one can do about that error during runtime (it
points to a broken installation), this should be an assert instead.

More specifically, such issue asserted before as well, but only on the

    CORRADE_INTERNAL_ASSERT(file);

line, which is totally unhelpful. Thanks to @AndySomogyi for pointing
out the issue and providing initial patch.