TinyGltfImporter: fix sampler state loading, adapt it to Magnum changes.
The minification/mipmap filter was not decombined and the wrapping modes
were in a wrong order. Also, since the generic values don't have the
same identifiers as GL anymore, we need to explicitly translate now,
instead of just casting.

Unfortunately there are no tests for this at the moment. Will be added
later.