AssimpImporter: handle the case of a single root node better.
With 3b67370ee71a451c70007fbe30fdc5d204299027 I made the imported object
count actually correspond to count of objects in the scene file. However
this breaks in case some postprocessing flags are present or when Assimp
for some reason just collapses everything into a root node. The
DART integration code is affected by this.

Now, in case there's just one root node with no children, I'm importing
it as a single object in the scene instead of ignoring it.