external: manually apply GCC 4.8 support patch on json.hpp.
This allows us to build the TinyGltfImporter on GCC 4.8. It was disabled
until now on Linux because the Travis builds were using GCC 4.7 and
json.hpp required 4.9, but now that we're dropping 4.7 support, it makes
sense to go an extra step and have this built under 4.8 as well.

Taken from https://github.com/nlohmann/json/pull/1257, applied the part
that touches single_include/nlohmann/json.hpp skipping the compiler
check on top (which doesn't apply) and doing the change by hand.

I could also update to json.hpp 3.3 / 3.4, but that version is 700 kB
(vs 480 kB for this one) and I have no reason to use that. I also hope
I'll never need to update this thing again.