{Assimp,OpenGex,TinyGltf}Importer: consistent external image loading.
Now it is failing with an error (and not an assert) if and only if
image is not embedded, the file is not opened from a filesystem and file
loading callbacks are not set. Previously it could errorneously assert
even when loading embedded images, try to look for images on the
filesystem when loading in-memory data or fail to open images when a
file inside current working directory was opened.