BasisImageConverter: reorganize the config.
Group it the same way as the basisu tool so it's easy to spot if there's
something missing (and for users easier to match the names), fix
spelling of some options and expand shorthands for others.

Also adding a bunch of missing options, while OTOH removing
read_source_images / write_output_basis_files as those make no sense to
have exposed.