BasisImageConverter: force the users to link to pthread.
After four hours lost trying to figure out a way around this *at
runtime*, I gave up -- linking the plugin to pthread (but not the app)
would cause the plugin to die on a null function pointer call and
there's no black magic that would allow me to detect that. So now the
plugin requires the app to link to pthread, otherwise failing to load
with pthread_create being an undefined symbol. That unfortunately breaks
it even for the single-threaded case, but what can we do :(