AssimpImporter: preserve transformation of the root node, if any.
In some cases (for example COLLADA files with Z_UP defined), the root
node contains a transformation that patches the up direction to be Y
again. Until now, AssimpImporter ignored this fact, causing meshes to be
imported in a wrong orientation; moreover this behavior was inconsistent
when PreTransformVertices got enabled.

This now preserved transformation of the root node and applies it to
direct children of the (ignored) root node, which both makes the
orientation correct again for Z-up files and also ensures the data
orientation is consistent with and without PreTransformVertices enabled.

Co-authored-by: Max Schwarz <max.schwarz@online.de>