StlImporter,StanfordImporter: avoid an extra copy when opening files.
With current importer interfaces we can't assume anything about lifetime
of views passed to doOpenData() and thus have to make a copy. But not
for files -- until now, this was taken care of by the base
implementation, which called into doOpenData() with full contents of the
loaded file (and then the plugins made a copy). With doOpenFile()
implemented on the plugin side we can avoid that copy, as we're owners
of the array and thus can avoid it going out of scope at the end of the
openFile() call.