PngImporter: brown magic.
magnum-plugins installed via homebrew fail to compile these two files
because FOR SOME REASON isnan is a macro. Impossible to replicate in a
normal, non-homebrew build. Wasted fifty years on this already, so I'm
trying this now. If that makes it work, great, I'll document this. If
not, it will get immediately reverted.