modules: properly set imported configurations for the Assimp target.
When IMPORTED_LOCATION_<CONFIG> is used instead of just IMPORTED_LOCATION
(which is the case when both debug and release library is found
instead of just one of them), CMake needs the IMPORTED_CONFIGURATIONS
populated as well so it knows which locations to pick. This wasn'ŧ done
and thus for CMake it looks like the IMPORTED target has no location,
but of ALL POSSIBILITIES how to handle this, CMake simply passes

    -lAssimp::Assimp

to the linker, with no message, no warning, fuck all. That's the same
behavior as from the wild 2.8.12 days, where instead of complaining that
a target (containing semicolons) didn't exist, it just passed it to the
linker and didn't care. Since 3.0 it properly warns (see the policy
https://cmake.org/cmake/help/v3.0/policy/CMP0028.html), so I was getting
mad why the target doesn't exist but CMake doesn't warn. Apparently
nobody thought about fixing this case as well. Oh CMake.

Crazily enough this problem was in the module since it was originally
contributed in 2017. Apparently nobody had a setup with both debug and
release binary present until now?