StbImageConverter: detect output format from filename.
Because I'm lazy and my own TgaImageConverter doesn't support RLE (which
I want to use for reducing size of shader test ground truth files in the
core Magnum repo) yet, I was always postprocessing the generated images
with StbImageConverter, as stb_image_write supports that. However doing
this:

    magnum-imageconverter --in-place -C StbImageConverter file.tga

always failed because StbImageConverter didn't know what file to use
because it didn't have a chance to see the filename. Silly limitation.