AssimpImporter: check we're not missing anything in the root scene.
Now that the ASS_DONT_DO_DUMB_SHIT flag is set, we can finally put an
assert for this. Well, except if the file is coming from outside, in
which case it's still full of shit, which requires us to be able to
detect such a case. On the other hand that means we don't have to create
an extra copy of the Assimp::Importer, which is nice since having just
one dumpster fire instead of two is always a good thing.