AssimpImporter: don't use APIs not present on builds without exports.
No idea why the C wrappers are not there but the underlying C++
implementation is. But it's assimp we're talking about here, any sanity
assumption go outta the window.

AND OF COURSE things are never simple with this pile of hacks, so we
still have to use the C wrappers for versions < 4.0. It also leaks
some allocated memory inside that CopyScene, so VERY WONDERFUL, YES.