*ImageConverter: don't attempt to convert zero-size/null images.
This is now asserted for by the base implementation. Result is a slight
reduction in test cases and for all plugins, in some cases (PNG, JPEG) I
had to invent new crazy ways to trigger a conversion failure now that I
can't abuse zero-sized images for that anymore.

In case of StbImageConverter there weren't any triggerable error paths
left, so I converted the runtime errors to assertions.