OpenExr{Importer,ImageConverter}: mipmap support.
Eh, wouldn't it be nice to have a library with a consistent and sane
access to both scanline and tiled files under ONE interface? This is
unnecessarily awful to deal with.

Also, why there's no sane way to detect incomplete tiles or at least
levels?! The amount of reverse engineering I had to do in order to punch
the info out of the API was ridiculous.