KtxImageConverter: test DFD against output of dfdutils.
Raw DFD data is bundled with the test, generated from a patch to a
dfdutils program, see README.md.

This removes some of the 2D compressed files as we only needed that many
to get some DFD test coverage, but we have a separate DFD test for all
formats now.