BasisImageConverter: support user-supplied mip levels
Unfortunately basisu doesn't flip level 1 or higher, so we have to do
that on our side.

If the mip_gen option is set to empty, it's determined from the number
of user-supplied levels -- if just one is passed, levels are generated,
otherwise the provided ones are used. If mip_gen is set to true and
custom levels are provided as well, it prints a warning.