BasisImporter: import video as separate images
Because transcoding all frames at once is probably a stupid idea. We
also catch any attempt to seek ahead as an error, since basis doesn't
allow seeking (all but the first image are P-frames).

There are no I-frame annotations in KTX2 + UASTC, so there is seeking
impossible, except to the first frame.