package/ci: download & build our own Zstd.
This sidesteps the single-file zstd bundled with Basis and all problems
it brings. It's still being used if Zstd is not found via any other
means in order to give existing packages (such as vcpkg) time to adapt,
but handling for it will eventually be removed, requiring the full
library always.