[wip] AssimpImporter: fix tests for 5.1.1
This needs minor changes to TinyGltfImporter test files, coming in the next commit. The last remaining failing test is an Assimp bug, fix follows in another commit.

We'll ignore 5.1.0 since there are only a few days in between and it's not worth special-casing the Collada light test because of this.