{Cgltf,TinyGltf}Importer: adapt tests to the new SceneData.
Reduced the test case for transformations as we don't handle object with
meshes in a separate branch from non-mesh objects; on the other hand
we're omitting fields in case they're not needed by any objects, so the
test got updated for this. Of course this uncovered a new tinygltf
misfeature/bug where TRS gets ignored if a matrix is present as well.
Sigh.

For multi-primitive meshes the importer-side handling is much simplified
as it's now only about duplicating the mesh field for given object, no
new objects being added to the hierarchy.