CgltfImporter: import material extensions handled by cgltf
Treat them like unknown extensions and import their attributes with the
original glTF names. Some of these may turn into a standard
MaterialAttribute or MaterialLayer in the future, but for the time being
this makes them accessible without changes to magnum. Tests will follow
in another commit.

There are no wrappers for these material types/layers, so we don't have
to test all these interactions that the clearcoat extension is going
through. Just checking the defaults and the texture import.