CgltfImporter: parse all custom material number primitives as float
Because consistently detecting if a number is int, unsigned int or float is impossible. Integer attributes may use exponent notation and decimal points followed by a 0, leaving us at the mercy of the glTF exporter used for the asset. Note that integer attributes in texture objects (texture index, coordinate set) are unaffected because we know the expected type.