CMake: explicitly set THREADS_PREFER_PTHREAD_FLAG.
I was aware that I was not doing this even though being explicitly told
to, but things worked so I didn't bother. But now, somehow, *one* CI job
started failing due to -lpthread doing nothing for some reason, and this
flag seems to fix that. So let's set that, and do it consistently
everywhere.

Cursed, my hands are cursed.