modules: Link to SpirvTools in FindGlslang.
That's the default build setup, and is what's used on Arch and Ubuntu.
On Arch it worked for me because there's a dynamic build (so the
transitive dependencies are resolved automatically), on the CI it
unfortunately worked also because I disabled SPIRV-Tools in the custom
Glslang build I have there, and that's apparently what Homebrew does as
well.

It didn't work with the new ARM64 build anymore, because there I take an
upstream Ubuntu package, which is a static library and which has the
SPIRV-Tools integration enabled.