CgltfImporter: adapt material tests to new guts.
The main behavioral change is that OOB errors don't fail the whole file,
but only a particular material, again with a better error message. Apart
from that, if a custom extension texture object has something wrong, it
gets ignored instead of failing the whole material -- after all, it
could be something completely different and only looking like a texture
reference.