Add back a deprecated CgltfImporter stub for backwards compat.
Implementation-wise it's just a typedef, for the user it's a plugin that
depends on GltfImporter. The only complexity in the stub is a copy of the
configuration, since that doesn't get shared from the base class, and
for that is also the only remaining test there.

Another option would be to keep CgltfImporter as a class to not make its
URL just disappear from the docs, but then I'd have to export
GltfImporter private symbols.

While it would be possible, I'm not adding deprecation warnings to CMake
if WITH_CGLTFIMPORTER is used. Since I need to build the plugin in many
places to verify it continues to work, it'd be too annoying.