{Cgltf,TinyGltf}Importer: skip object ID when registering custom attribs.
Possibly an oversight carried over to the new implementation, or maybe a
leftover from the time when object ID wasn't a builtin attribute yet.
Since the attribute gets reported as MeshAttribute::ObjectId, it doesn't
make sense to expose it through meshAttribute[For]Name(), the identifier
from there won't ever be used in anything.