CgltfImporter: test a corner case with view smaller than count*stride.
This is valid glTF, but our StridedArrayView asserts expect the input
view to contain all count*stride bytes (which is reasonable given
there's no way to know how much of the view is going to be actually
used).

The better way would be to form and then shift the view instead of
faking the size, but that'd need to figure out some
StridedArrayView::shift() API first. I have more than enough TODOs
already, so this has to wait.