{Cgltf,TinyGltf}Importer: merge mesh OOB + type tests back with invalid.
In this case I made an "archived" copy of the test file for TinyGltf,
because it's diverging enough already -- a dedicated test case for
doubles which we don't care about, a test for OOB index accessor that
has to be separate, etc. Additionally I removed the binary buffer and
replaced that with embedded zeros, because we're not testing the
contents anywhere anyway.