{Cgltf,TinyGltf}Importer: rework an excessively awful texture test.
Sure, exporting a file from Blender with everything including a light
and a camera. Way to go!

This also means I can merge the other two files into it, which were
presumably created only because the texture.gltf was just way too
complex to be patched by hand.

And the binary variants are also useless, since textures on its own
don't contain anything that would need to be fetched from a buffer.