KtxImageConverter: warn when no orientation is set.
Since KtxImporter warns when the KTXorientation key isn't present, it
makes sense to do the same in the converter as well, and doing so will
uncover nasty issues when a completely empty configuration is passed to
the plugin by accident.

This change is further complicated by the fact that our ground truth
test data are generated with toktx, and KTX Software 4.0 has a bug where
it doesn't write KTXorientation for 1D images. That seems to be fixed in
4.1 RC 1, but that version writes a lot longer KTXwriter value into the
file, so it feels weird to just regenerate all files with it. Instead,
I'm faking the 1D files for now by directly taking the output of
KtxImageConverter, and will regenerate them all using toktx once 4.1 is
out.

A much better solution would be to have toktx not write any writer
metadata to the file so the ground truth data wouldn't all need to be
regenerated for a new version, but that's apparently impossible unless I
submit a patch, I guess.