DdsImporter: actually robust handling of short files.
It was just happily reading the first four bytes, even though the file
could as well be empty. That was fortunately asserting, so not a nasty
crash or something worse, but then the file header was also happily read
without checking the size. Uh oh.

Also improved the messages to contain the actually important information
and not the unimportant information (who cares if some fourcc is set).

What I find especially funny is that Microsoft's documentation says the
file has a 4-byte signature and then a 124-bit header. I wonder what was
the thought process there -- if I'm about to detect what type is the
file, then I probably need just those four bytes (or any other prefix),
and if I'm loading a DDS file then I need all 128 anyway, so why not
have the DDS header as a single piece. Or is that another historical
reason I'm not aware of? The header acting also as a D3DSURFACE or some
such??