DdsImporter: add comments for fields that have fixed contents.
Not sure if I should check this. Probably not. If these are wrong, then
the rest of the file might as well be just fine and it's just the
exporter that didn't bother; then it would be just a noisy warning or a
useless error. And if the file is broken completely, then I can't really
detect that by checking just these fields.