DdsImporter: rework for proper 1D, cube and array support.
What was there was ... not working at all, mixing array slices and mip
levels together, not importing cubemaps as 3D, .... I suppose that was
because the code originated at a time where the importer interfaces had
no support for mip levels, or the GL library no support for uploading
cubemaps as a single 3D image, and so there wasn't really any other way
to implement this.

Also further tightening up various checks, like disallowing depths for
non-3D textures, or failing if both 3D and cubemap flag is set.