DdsImporter: implement Y- and Z-flipping.
Consistently with basically all other image importers. This was a
serious omission, however given the file format is most often used for
block-compressed data (where we can't flip as easily yet), it was not
*that* critical.

The behavior is mostly consistent with what got invented by pezcode for
KtxImporter. I just shamelessly compied most of the logic from there.