GltfImporter: reformat test glTF files with two-space indent.
For consistency with the upcoming glTF exporter and less waste, as I'll
be adding quite a few new extensions now. Done mainly with

    python -m json.tool <file>.gltf{,} --indent 2

but *of course* it doesn't just reformat, it also changes Unicode
characters to escapes, drops duplicate attributes we're explicitly
testing for, etc. So I had to use

    git diff -w

and go over the changes until they looked innocent. Due to that, this
commit also contains just the formatting changes, but not adaptation of
file/line info in the tests.