GltfImporter: make material texture specification less error-prone.
Having to repeat a very similar attribute string three times for
questionable reasons (saving some temporary allocations or string
copies?) is only asking for trouble. Originally this was an enum, which
wasn't any better, one can do a copy-paste mistake in either, but with
introduction of custom attributes and the upcoming texture layer
attributes this is just too much to handle.