OpenExrImageConverter: recognize and rely on ImageFlag3D::CubeMap.
This is a breaking change that obsoletes the envmap=cube option for 3D
images. Instead it's required to be always empty for 3D and the image
has to contain the CubeMap flag instead. Additionally, the converter
checks and fails if CubeMap|Array is set, and suggests exporting each
separately.

I don't expect that anyone was seriously using this option so I didn't
bother with providing backwards compat. Let me know if you need that,
then I'll add that back.