BasisImageConverter: recognize ImageFlag*D::Array and CubeMap.
This also means the plugin no longer needs to unconditionally warn that
a 3D image is being exported as 2D array, it does so only if
ImageFlag3D::Array is not present. The test was updated to pass the
Array flag where possible, yet still verifying that the flag is
correctly printed in the 3D case.

Furthermore, as Basis compresses in 4x4 blocks, 1D array images are now
explicitly disallowed -- the layer cross-talk would make them unusable
anyway.