AstcImporter: mark 3D images with 2D compression as 2D arrays.
I thought about providing some "assumeCubeMap", "assume3D" etc. options
to provide better control over this as the format is amazingly low on
metadata, but postponing that to once a practical use case gets found.

So far it doesn't seem that the ARM ASTC Encoder is even capable of
dealing with things like arrays or cubemaps, so this feature is really
just for completeness.