GltfImporter: allow images interleaved with other data in glb2gltf.py.
Not really sure what the assert was about. Possibly to check that no
other view overlaps with the to-be-removed buffer view containing an
image? But even that would work, since the data are composed from all
other buffer views instead of cutting away buffer views that belong to
an image.