GltfImporter: use the easier & less error-prone material comparator.
Going through the public material interface had the disadvantage of not
showing what attributes are actually present and what not. I.e., some
could be missing altogether yet they be present but the implicit default
papered over that; or hiding the difference between packed and separate
textures...

And on the other hand the interface made some things harder, such as
having to test for texture presence first. Which is fine (and desired)
for actual app usage, but not for test when we want a 1:1 comparison.