[Tiny]GltfImporter: deconfuse texture coordinate Y-flip test files.
Not sure what was I doing there, but on the first look it wasn't clear
at all why the test file has to be separated in half. Now I know, it's
because textureCoordinateYFlipInMaterial would be enabled implicitly for
one and not the other, but that should have been clearly mentioned in a
comment.

Furthermore, some of the meshes were named as "integer", implying a
32-bit integer, while that was not really the case. I probably meant to
imply "any integer", but just saying "byte" and "short" would imply that
too.

Also dropping the "material" from the test case name, since it involves
meshes as well.