AssimpImporter: skip skinning attribute related tests on 5.2.4
It's completely broken, reporting wrong attribute count or just wrong
attribute data. For some reason it works with glTF files but that might
just be a weird interaction with the other brokenness there. It got
fixed right away in 5.2.5, and the importer can't really detect or
handle it anyway, so no change there.