GltfSceneConverter: properly fill extensionsUsed/Required for images.
I was thinking about the extensions so much that I actually forgot to
add them to extensionsUsed and extensionsRequired, sigh. Only discovered
this while writing tests for magnum-sceneconverter.

Also I realized it makes sense to fill extensionsUsed if the image is
added, to hint that the image might need some special handling. However,
because the image itself doesn't use any extension object, it's
impossible to tell which image uses which extension. The full
information is there only once the image gets referenced by a texture
which actually uses an extension object, and which triggers the
extension to be added to extensionsRequired.

Finally, the extensionsUsed being properly filled now uncovered a corner
case where an image adding failure would corrupt the internal state in a
way that makes it impossible to add any images after. Fixed that with a
bit of code shuffling and hardened tests.