AssimpImporter: don't use a growable array for indices.
There was already an assert checking that each face has no more than 3
indices. But since a two-/one-vertex triangle or a one-vertex line
doesn't make sense, we can just check for equality instead. And then it
means the final index array size can be known in advance. The check is
tighter now, but assuming implementation sanity (*cough*) it shouldn't
cause random models to suddenly start asserting.

Not using a growable array means that this fixes a particularly annoying
case happening when static plugins are linked to more than one dynamic
library, causing the growable deleter pointer to be duplicated and
resulting in the following assertion:

    Trade::AbstractImporter::mesh(): implementation is not allowed to use a custom Array deleter

The case of static plugins, multiple dynamic libraries and custom
deleters needs to be fixed in a more general way, but for now this
should suffice.