StbResizeImageConverter: add an ability to disable upsampling.
I realized that adding a simple boolean toggle for this is the easiest
possible solution. The original idea was to introduce a new `sizeMax`
option that would resize to that size or smaller, but it felt
overcomplicated and I put that aside.

Now it also performs a direct copy if the target image size is exactly
the same as the input regardless of whether upsampling is disabled,
which seems like a reasonable thing to do. Without that, it would
perform a "seemingly no-op" resampling, which is both slower and
introduces artifacts, apparently.