GltfImporter: import invalid vertex formats as custom attributes.
Instead of failing the whole mesh import. Because it can happen that a
widely used dataset has some minor bug in their files (such as vertex
colors without the normalized bit) and forcing users to hand-patch the
glTFs isn't feasible. Importing them as custom attributes instead makes
it possible to automate the fixing.

This also significantly reduces the amount of "invalid mesh" cases that
need to be tested. Frankly, it probably didn't need to be this complex
in the first place, and it still didn't check that we're not
accidentally allowing invalid types (because that isn't really
possible, apart from generating files that enumerate all possible
variants). Coverage is still at 100% even with all those cases removed.

The original behavior can be restored with the recently added strict
option.