GltfImporter: use the builtin MeshAttribute::JointIds and Weights.
The old custom "JOINTS" and "WEIGHTS" attributes are kept for backwards
compatibility and can be turned off if desired.

A GltfSceneConverter test relied on these being imported as custom, so
it's patched to have the compatibility disabled. Update to
GltfSceneConverter itself to recognize the builtin attributes will
follow eventually as well.