AssimpImporter: use the builtin MeshAttribute::JointIds and Weights.
Same as with GltfImporter, the old custom "JOINTS" and "WEIGHTS"
attributes are kept. The difference is that now Assimp can have the
array size arbitrary, ranging from 1 to however many (in contrast to
glTF, where it's always a multiple of four).

Another difference is that maxJointWeights is no longer 4 by default,
but set implicitly to 0, i.e. allowing any number of joint weights. This
makes it consistent with GltfImporter, where it's not doing any
postprocessing by default either.

If the end-user app needs to impose some limit on the weight count, or
have it an exact multiply of four for example, that is something that
should be handled by some MeshTools utility, not the plugin itself. I
still have to write those though, together with some quantization /
packing utility.