PngImporter: drop manual file signature comparison.
Turns out none of this is needed, as the library can do it on its own,
and produce nicer error messages. The code here was even telling it to
"assume the signature was already compared", sigh.

I wonder where did I copy all this insanity from, the usage is far
simpler than it seemed at first.