[wip] Png{Importer,ImageConverter}: attempt to work around longjmp issues.
This makes the code no longer use png_jmpbuf() in an attempt to fix the
clang-cl error, as I suspected the misbehavior was due to longjmp being
possibly called even before setjmp() is called in our code.

But, going through libpng sources, it seems that for the version
mismatch check the png_create_*_struct() itself calls setjmp(), so there
should be no problem, and this whole extra complication is not needed.

And, indeed, it has no effect on the behavior -- the clang-cl build
still crashes or asserts due to some stack corruption.