Png{Importer,ImageConverter}: don't assert on libpng patch ver changes.
Just a temporary hotfix, until a robust solution is found that works on
all compilers -- the proper fix would be to set up the error callbacks
directly in the png_create_*() APIs, which would then catch also the
version mismatches, but all my attempts to do that run into *nasty*
issues with setjmp/longjmp on a single version of clang-cl. Will push
that to a branch / PR for further investigation, this is just to unblock
people that had libpng updated recently from 1.6.38 to 39.