UfbxImporter: don't unconditionally generate a trivial index buffer.
If generateIndices is disabled, one would expect that an index buffer
isn't present, which wasn't the case for some reason. The
removeDuplicates() API doesn't need to have the input indexed either (in
fact, if the mesh is indexed already, it means one extra step), so just
delete all that altogether. I suspect this was a leftover from when the
code originally used ufbx's own deduplication API.

Furthermore, the generateIndices option wasn't tested at all, so take
this as a chance to verify both scenarios. Fortunately (or
unfortunately?) the mesh test cases have the input files without any
duplicate data, so the only difference is that the index buffer is or
isn't present, and if present it's trivial.