UfbxImporter: expose no scene for *.mtl files.
Being able to open *.mtl files directly is a very nice feature. If
somebody does that explicitly, it's most probably to parse a material
library -- and then it doens't make sense to expose an empty scene. Same
behavior is with glTFs containing just "materials" entries and no
"nodes" or "scenes".

(For completeness, AssimpImporter tries to open *.mtl files as FBX. Not
useful.)