GltfSceneConverter: check texture ID and layer for all material attribs.
Maintaining a "whitelist" of texture attributes proved to be extremely
error-prone even after adding the first ever extension and the only way
to ensure the whitelist is complete would be to write a similarly
convoluted test case, *and then* somehow ensuring the test case isn't
missing any attributes.

So it's now going through all material attributes instead and picking
all that look like are a texture and are builtin. The assumption here is
that eventually the converter plugin will support all builtin material
attributes, so it doesn't make sense to do it any other way, and if
there would be some rare exceptions, they can be added as special-cases
(such as is now for the Phong DiffuseColorTexture, which is
builtin but ignored so it doesn't need to be bounds-checked).