GltfImporter: rethink behavior with empty node extras arrays.
Warning for them doesn't make sense, since even if their type was
fully known in advance they wouldn't contribute to the output in any
way. So instead they're just ignored.

Ideally their name would be registered with a placeholder type first and
then the type would get concretized once a non-empty or non-array value
is found, but that's additional non-trivial logic and testing. I'm
adding just a TODO for now.