GltfImporter: spot lights with a 90° outer angle should be allowed.
Used >= instead of > by accident. Also, since the number parsing
and import is all with 32-bit floats while the native JSON number type
is 64-bit floats I extended the tests to verify that double-precision
half pi is properly accepted, without some rounding somewhere causing
it to not be.

To further prevent rounding errors, the importer code is now also using
the constants directly instead of relying on degree-to-radian conversion.

The previous light.gltf file is copied to TinyGltfImporter in order to
not have to deal with the deprecated plugin as well.