StbImageConverter: compare the output HDR files to a ground truth.
I need to be able to easily regenerate those file, which is the main
reason. However it turns out that the file used by StbImageImporter was
the worst of them all -- it was apparently the output of the hdrRg()
test, which just duplicates one channel three times. I renamed that to
rrr.hdr and made a new rgb.hdr that is actually different channels.