GltfImporter: tighten up array parsing in scene node extras.
Previously, if there was an array, it was treated the same way as if
there were multiple keys with the same value, and having a single key
once a scalar and another time an array was accepted. This however made
it impossible to know whether a particular custom scene field is meant
to be an array or not when exporting a glTF back.

To fix that, if an extra key is first found as an array, it's marked
with SceneFieldFlag::MultiEntry, which hints that it should get exported
as an array again. If the same key is then found as a scalar, it's
ignored with a warning. And conversely, if an extra key is first found
as a scalar, finding it later as an array will produce a warning. This
way it's actually possible to have the extras rountrip correctly on
export.

This is breaking change in behavior, however I don't think there were
any widely used assets that would mix up arrays and scalars like this.
Even the most typeless use cases still want some consistency.