GltfSceneConverter: shy people away from [imageConverter] options.
This was more of a hack than a real feature in order to allow the
options to be set from the magnum-sceneconverter interface. That utility
has now a --set option that can set the options globally for the
image converter plugin, or any other plugin as well.

I'm not deprecating this just yet because a lot of existing batch
scripts relies on this working. But it'll get removed eventually, yes.