GltfSceneConverter: write min/max properties for POSITION accessors.
As required by the spec. Same as with the target property, my opinion is
that validating the correctness has a significantly higher cost than
just calculating the min/max on the fly during import -- because in
order to be sure you can rely on those values, you *still* have to
calculate the min/max anyway.

Also, what use is for these? What if I don't have an AABB-based culling
but sphere-based? Where is such info meant to be stored?