DdsImporter: Y-flip BC1/2/3/4/5 blocks on load.
Which means it'll no longer warn for them. Can be disabled with the
assumeYUpZBackward option, which -- I assume -- was what people already
did before if the data were loading fine for them but they were annoyed
by the warning.

Also, two new Compressonator invocations, two new bugs. If I ever get an
idea to integrate this library as a plugin, smash me with a brick.